Ticket #10 - Code cleanup.

+++ TravellingSalesman/Utils/CenteringConverter.cs
@@ -4,11 +4,15 @@ using System.Windows.Data;
 
 namespace TravellingSalesman.Utils
 {
+    /// <summary>
+    /// This converter is responsible to center the drawn ellipses.
+    /// Because normally they would be drawn on the top left corner.
+    /// </summary>
     public class CenteringConverter : IValueConverter
     {
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            if (value is int && parameter is string)
+            if (value is double && parameter is string)
             {
                 int offset;
                 if (!int.TryParse((string)parameter, out offset))
@@ -16,7 +20,10 @@ namespace TravellingSalesman.Utils
                     offset = 0;
                 }
 
-                return (int)value - (offset / 2.0);
+                // The offset is the width/height of the ellipse. 
+                // Because height and width are the same for this ellipse
+                // only one offset is needed.
+                return (int)((double)value - (offset / 2.0));
 
             }
             return value;
